# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Real Estate Operations Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: assignedAgentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: dueDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── APPOINTMENTS (CALENDAR) ─────────────────────────────────────────────────
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List all automation workflows
      responses:
        "200":
          description: Workflow list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a workflow rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          description: Not found
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workflows/{id}/toggle:
    patch:
      operationId: toggleWorkflow
      tags: [workflows]
      summary: Toggle workflow active status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}/executions:
    get:
      operationId: getWorkflowExecutions
      tags: [workflows]
      summary: Get execution history for a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Executions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowExecution"
  /workflow-executions:
    get:
      operationId: listWorkflowExecutions
      tags: [workflows]
      summary: List all workflow executions
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Execution list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowExecution"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, type, status, priority, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        type: { type: string }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: ["string", "null"] }
        dueTime: { type: ["string", "null"] }
        assignedAgentId: { type: ["integer", "null"] }
        assignedAgentName: { type: ["string", "null"] }
        leadId: { type: ["integer", "null"] }
        clientId: { type: ["integer", "null"] }
        createdAt: { type: string }
    TaskInput:
      type: object
      required: [title, type]
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: string }
        dueTime: { type: string }
        assignedAgentId: { type: integer }
        leadId: { type: integer }
        clientId: { type: integer }
    TaskUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        status: { type: string }
        priority: { type: string }
        dueDate: { type: ["string", "null"] }
        dueTime: { type: ["string", "null"] }
        assignedAgentId: { type: ["integer", "null"] }
        leadId: { type: ["integer", "null"] }
        clientId: { type: ["integer", "null"] }

    # ─── APPOINTMENT ─────────────────────────────────────────────────────────────
...
    Workflow:
      type: object
      required: [id, name, triggerEvent, isActive, executionCount, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        triggerEvent: { type: string }
        conditions: { type: ["string", "null"] }
        actions: { type: string }
        isActive: { type: boolean }
        executionCount: { type: integer }
        lastExecutedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    WorkflowInput:
      type: object
      required: [name, triggerEvent, actions]
      properties:
        name: { type: string }
        description: { type: string }
        triggerEvent: { type: string }
        conditions: { type: string }
        actions: { type: string }
        isActive: { type: boolean }
    WorkflowUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        triggerEvent: { type: string }
        conditions: { type: string }
        actions: { type: string }
        isActive: { type: boolean }
    WorkflowExecution:
      type: object
      required: [id, workflowId, workflowName, status, triggeredBy, executedAt]
      properties:
        id: { type: integer }
        workflowId: { type: integer }
        workflowName: { type: string }
        status: { type: string }
        triggeredBy: { type: string }
        triggerEntityId: { type: ["integer", "null"] }
        actionsExecuted: { type: integer }
        errorMessage: { type: ["string", "null"] }
        executedAt: { type: string }
...